Preserve unbacked SymInt on SymNode
Previously, when we applied a replacement, a SymInt that was
previously an unbacked SymInt would then transmute into whatever
we replaced it into (e.g., a constant).

This has a major downside: we often look at SymInts associated with
FX nodes (e.g., the meta of x.item() return) to find out where the
unbacked SymInt was allocated.  If we replace it, we no longer can
find out where, e.g., u1 was allocated!  But we need to know this
so we can generate deferred runtime asserts like u1 == s0.

To solve this problem, I separate out unbacked replacements into a
separate dict unbacked_replacements and avoid applying them when you
query for the sympy.Expr associated with a SymNode.  Everything else
still respects these; importantly, if you want to check if u1 statically
equals s0, we will still end up doing the replacement.

I'm open to other ways to solve this problem, but this one didn't seem
too hard to implement.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: 80e339ba4b53bbadb49b3b3127aa112528676340
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120816